refactor: consolidate container features into single service
  - Merge EnvironmentFeature, LogsFeature, PortsFeature, VolumesFeature into ContainerService
  - Remove unnecessary abstraction layers for simple data extraction methods
  - Update aws_service.py and ContainerUI to use consolidated service
  - Delete 4 feature files and empty features directory
  - Reduce codebase by ~200 lines while maintaining all functionality
